Asking the users if they are sure

+++ encryptor.py
@@ -8,17 +8,21 @@ for file in os.listdir():
 
     files.append(fileName)
     break
-
+    
 print(files)
 
-key = Fernet.generate_key()
+confirmationMessage = input("Are you sure you want to encrypt this file(y/N)?: ")
 
-with open("decryption_key.txt", "wb") as decryption_key:
-    decryption_key.write(key)
+if confirmationMessage == "Y" or "y":
+    
+    key = Fernet.generate_key()
 
-for file in files:
-    with open(file, "rb") as thefile:
-        contents = thefile.read()
-    contents_encrypted = Fernet(key).encrypt(contents)
-    with open(file, "wb") as thefile:
-        thefile.write(contents_encrypted)
+    with open("decryption_key.txt", "wb") as decryption_key:
+        decryption_key.write(key)
+
+    for file in files:
+        with open(file, "rb") as thefile:
+            contents = thefile.read()
+        contents_encrypted = Fernet(key).encrypt(contents)
+        with open(file, "wb") as thefile:
+            thefile.write(contents_encrypted)
